Refactor shared test utility methods for loading fixtures
