+++ test/index.html
@@ -13,6 +13,9 @@
     <script src="/vendor/mocha.js"></script>
     <script src="/vendor/chai.js"></script>
     <script src="/vendor/sinon-4.1.6.js"></script>
+    <!-- local test utils -->
+    <script src="utils.js"></script>
+
     <script>mocha.setup('bdd')</script>
 
     <!-- load source code to be tested -->
